+++ app/src/main/java/ru/rtuitlab/itlab/domain/use_cases/events/GetEventsUseCase.kt
@@ -2,8 +2,8 @@ package ru.rtuitlab.itlab.domain.use_cases.events
 
 import kotlinx.coroutines.flow.map
 import ru.rtuitlab.itlab.domain.repository.EventsRepository
-import ru.rtuitlab.itlab.common.endOfTimes
-import ru.rtuitlab.itlab.common.nowAsIso8601
+import ru.rtuitlab.itlab.common.extensions.endOfTimes
+import ru.rtuitlab.itlab.common.extensions.nowAsIso8601
 import javax.inject.Inject
 
 class GetEventsUseCase @Inject constructor(
