+++ app/src/main/java/ru/rtuitlab/itlab/domain/use_cases/events/UpdateEventsUseCase.kt
@@ -3,7 +3,7 @@ package ru.rtuitlab.itlab.domain.use_cases.events
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import ru.rtuitlab.itlab.domain.repository.EventsRepository
-import ru.rtuitlab.itlab.common.nowAsIso8601
+import ru.rtuitlab.itlab.common.extensions.nowAsIso8601
 import javax.inject.Inject
 
 class UpdateEventsUseCase @Inject constructor(
