+++ app/src/main/java/ru/rtuitlab/itlab/presentation/UserViewModel.kt
@@ -7,6 +7,8 @@ import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
 import kotlinx.datetime.DateTimeUnit
 import ru.rtuitlab.itlab.common.*
+import ru.rtuitlab.itlab.common.extensions.minus
+import ru.rtuitlab.itlab.common.extensions.toIsoString
 import ru.rtuitlab.itlab.domain.use_cases.users.GetUserPropertyTypesUseCase
 import ru.rtuitlab.itlab.domain.use_cases.users.GetUserUseCase
 import ru.rtuitlab.itlab.domain.use_cases.events.GetUserEventsUseCase
